backend/requirements.txt
# fastapi
# uvicorn[standard]
# psycopg2-binary
# SQLAlchemy
# python-dotenv
# passlib[bcrypt]
# python-jose[cryptography]

fastapi
uvicorn[standard]
sqlalchemy
pymysql          # MySQL driver
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
